Update Main.cpp
adds decimal support for mean only wierdly

+++ Main.cpp
@@ -1,17 +1,20 @@
+#include <string>
 #include <iostream>
+#include <iomanip>
+
 
 int askNumber()
 {
 	std::cout << "enter integer: ";
-	int input{}; 
+	long double input{};
 	std::cin >> input; // Gets input
 	return input;
 }
 
-int getTotal_andMean(int a, int b, int c, int d, int e) // Gets the Total and Mean of the 5 integers
+int getTotal_andMean(long double a, long double b, long double c, long double d, long double e) // Gets the Total and Mean of the 5 integers
 {
-	int x{ a + b + c + d + e }; // Adds them up
-	int y{ x / 5 };				// Divides total to 5
+	long double x{ a + b + c + d + e }; // Adds them up
+	long double y{ x / 5 };				// Divides total to 5
 	std::cout << "The total is: " << x << '\n';
 	std::cout << "the mean is: " << y << '\n';
 	return x;
@@ -25,5 +28,5 @@ int main()
 	int d{ askNumber() };
 	int e{ askNumber() };
 	getTotal_andMean(a, b, c, d, e);
-	return 0;	
+	return 0;
 }
